Updated plugin to support Gradle 8.x+
Refactored extension DSL property to task property mapping approach to
better align with Gradle-recommended approach. Updated package naming to
align with new governance structure - still pending input from Gradle
Plugin Portal engineers on updating group ID.
