+++ build.gradle
@@ -1,34 +1,30 @@
 plugins {
-	id 'java-gradle-plugin'
-	id 'com.gradle.plugin-publish' version '0.11.0'
+	id 'com.gradle.plugin-publish' version '1.1.0'
 }
 
 repositories {
 	mavenCentral()
-	jcenter()
 }
 
 dependencies {
 	implementation 'org.apache.commons:commons-lang3:3.9'
-	implementation 'commons-io:commons-io:2.6'
-	implementation 'org.apache.velocity:velocity-engine-core:2.2'
+	implementation 'commons-io:commons-io:2.7'
+	implementation 'org.apache.velocity:velocity-engine-core:2.3'
 }
 
-pluginBundle {
-	website = 'https://bitbucket.org/cpointe/flyway-java-migration-checksum-gradle-plugin'
-	vcsUrl = 'https://bitbucket.org/cpointe/flyway-java-migration-checksum-gradle-plugin.git'
-	tags = ['flyway']
-}
-
-version = '1.0.0'
+group = 'org.bitbucket.cpointe'
+version = '2.0.0'
 
 gradlePlugin {
+	website = 'https://github.com/TechnologyBrewery/flyway-java-migration-checksum-gradle-plugin'
+	vcsUrl = 'https://github.com/TechnologyBrewery/flyway-java-migration-checksum-gradle-plugin.git'
 	plugins {
 		flywayJavaMigrationChecksumPlugin {
 			id = 'org.bitbucket.cpointe.flyway-java-migration-checksum-gradle-plugin'
 			displayName = 'Java-based Flyway Migration Checksum Generation Plugin'
 			description = 'A plugin that provides baseline approach for generating checksums for Java-based Flyway database migrations'
-			implementationClass = 'org.bitbucket.cpointe.FlywayJavaMigrationChecksumPlugin'
+			tags.set(['flyway'])
+			implementationClass = 'org.technologybrewery.FlywayJavaMigrationChecksumPlugin'
 		}
 	}
 }
